+++ ocaml-lsp-server/test/e2e-new/test.ml
@@ -103,8 +103,13 @@ end
 
 open Import
 
-let start_client ?(capabilities = ClientCapabilities.create ()) ?workspaceFolders client =
-  Client.start client (InitializeParams.create ~capabilities ?workspaceFolders ())
+let start_client
+      ?(capabilities = ClientCapabilities.create ())
+      ?workspaceFolders
+      ?trace
+      client
+  =
+  Client.start client (InitializeParams.create ~capabilities ?workspaceFolders ?trace ())
 ;;
 
 let shutdown_client client =
@@ -146,6 +151,7 @@ module T : sig
     -> ?timeout:float
     -> ?capabilities:ClientCapabilities.t
     -> ?workspaceFolders:WorkspaceFolder.t list option
+    -> ?trace:TraceValue.t
     -> (unit Client.t -> 'a Fiber.t)
     -> 'a
 end = struct
@@ -234,11 +240,12 @@ end = struct
         ?timeout
         ?(capabilities = ClientCapabilities.create ())
         ?workspaceFolders
+        ?trace
         f
     =
     run ?extra_env ?handler ?stderr ?timeout
     @@ fun client ->
-    let run_client () = start_client ~capabilities ?workspaceFolders client in
+    let run_client () = start_client ~capabilities ?workspaceFolders ?trace client in
     let run_test () =
       let* (_ : InitializeResult.t) = Client.initialized client in
       f client